requirements.txt
# =========================
# Already installed
# =========================
# crewai
# google-genai

# =========================
# Needed for FastAPI backend
# =========================
fastapi
uvicorn[standard]
pydantic

# =========================
# Useful/required for env + HTTP calls
# =========================
python-dotenv
requests

# =========================
# Serper / tools
# =========================
crewai-tools

# =========================
# Markdown + PDF export
# =========================
markdown
weasyprint